+++ src/main/java/group/msg/jpowermonitor/config/CsvColumn.java
@@ -2,8 +2,6 @@ package group.msg.jpowermonitor.config;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 /**
  * Data class for csv column config.
  *
@@ -13,5 +11,5 @@ import java.math.BigDecimal;
 public class CsvColumn {
     private int index;
     private String name;
-    private BigDecimal energyInIdleMode;
+    private Double energyInIdleMode;
 }
