+++ src/main/java/group/msg/jpowermonitor/dto/DataPoint.java
@@ -3,7 +3,6 @@ package group.msg.jpowermonitor.dto;
 import group.msg.jpowermonitor.agent.Unit;
 import lombok.Value;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -12,7 +11,7 @@ import java.time.LocalDateTime;
 @Value
 public class DataPoint implements PowerQuestionable {
     String name;
-    BigDecimal value;
+    Double value;
     Unit unit;
     LocalDateTime time;
     String threadName;
